Confirm separate task file results
